Break apart single module into multiple
